+++ packages/plugin-web-attribution-browser/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.14](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-web-attribution-browser@2.2.13...@amplitude/plugin-web-attribution-browser@2.2.14) (2026-06-25)
+
+**Note:** Version bump only for package @amplitude/plugin-web-attribution-browser
+
+
+
+
+
 ## [2.2.13](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-web-attribution-browser@2.2.12...@amplitude/plugin-web-attribution-browser@2.2.13) (2026-06-11)
 
 **Note:** Version bump only for package @amplitude/plugin-web-attribution-browser